Объединение классов Extension и Provider. Удаление лишних глобальных свойств
